+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode/
 build/
-CMakeFiles/
+CMakeFiles
+Testing
