+++ examples/serialization/main.cpp
@@ -2,7 +2,7 @@
 #include "CNtity/Helper.hpp"
 
 #include <iostream>
-#include <map>
+#include <sstream>
 
 ////////////////////////////////////////////////////////////
 struct Position
