+++ tests/include/Components.hpp
@@ -1,17 +1,20 @@
 #pragma once
 
+////////////////////////////////////////////////////////////
 struct Position
 {
     int  x{}, y{};
     bool operator==(const Position& other) const { return x == other.x && y == other.y; }
 };
 
+////////////////////////////////////////////////////////////
 struct Velocity
 {
     float vx{}, vy{};
     bool  operator==(const Velocity& other) const { return vx == other.vx && vy == other.vy; }
 };
 
+////////////////////////////////////////////////////////////
 struct Health
 {
     int  hp{};
